+++ src/dashboardWebView/utils/index.ts
@@ -3,4 +3,5 @@ export * from './darkenColor';
 export * from './getRelPath';
 export * from './opacityColor';
 export * from './preserveColor';
+export * from './sorting';
 export * from './updateCssVariables';